Fix harmless compiler warnings.
FossilOrigin-Name: 0a926994691ccb02ea4f0ba3c2e8f728871ae67d24c705fbc0f7af81f38d5f47

+++ ext/session/sqlite3session.c
@@ -1102,7 +1102,7 @@ static int sessionTableInfo(
   
     i = 0;
     if( bRowid ){
-      int nName = strlen(SESSIONS_ROWID);
+      size_t nName = strlen(SESSIONS_ROWID);
       memcpy(pAlloc, SESSIONS_ROWID, nName+1);
       azCol[i] = (char*)pAlloc;
       pAlloc += nName+1;
